アーキテクチャファイルテンプレート修正
