+++ EngineerMemo.xctemplate/My MVVM Architecture/ContentView.xctemplate/___FILEBASENAME___.swift
