+++ EngineerMemo.xctemplate/My MVVM Architecture/Routing.xctemplate/TemplateInfo.plist
@@ -2,10 +2,6 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>HiddenFromLibrary</key>
-	<false/>
-	<key>HiddenFromChooser</key>
-	<false/>
 	<key>Kind</key>
 	<string>Xcode.IDEFoundation.TextSubstitutionFileTemplateKind</string>
 	<key>AllowedTypes</key>
@@ -13,10 +9,8 @@
 		<string>public.swift-source</string>
 	</array>
 	<key>DefaultCompletionName</key>
-	<string>~~~ContentView</string>
-	<key>MainTemplateFile</key>
-	<string>___FILEBASENAME___.swift</string>
+	<string>Routing</string>
 	<key>SortOrder</key>
-	<string>10</string>
+	<string>1</string>
 </dict>
 </plist>
