+++ EngineerMemo.xctemplate/My MVVM Architecture/ViewModel.xctemplate/TemplateInfo.plist
@@ -2,10 +2,6 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>HiddenFromLibrary</key>
-	<false/>
-	<key>HiddenFromChooser</key>
-	<false/>
 	<key>Kind</key>
 	<string>Xcode.IDEFoundation.TextSubstitutionFileTemplateKind</string>
 	<key>AllowedTypes</key>
@@ -13,10 +9,8 @@
 		<string>public.swift-source</string>
 	</array>
 	<key>DefaultCompletionName</key>
-	<string>~~~ViewModel</string>
-	<key>MainTemplateFile</key>
-	<string>___FILEBASENAME___.swift</string>
+	<string>ViewModel</string>
 	<key>SortOrder</key>
-	<string>40</string>
+	<string>4</string>
 </dict>
 </plist>
