+++ Scripts/Template/template.sh
@@ -1,7 +1,7 @@
-TEMPLATE_DIR=~/Library/Developer/Xcode/Templates/File\ Templates/MVVM
+TEMPLATE_DIR=~/Library/Developer/Xcode/Templates/File\ Templates
 
 if [ ! -d "${TEMPLATE_DIR}" ]; then
     mkdir -p "${TEMPLATE_DIR}"
 fi
 
-cp -r EngineerMemo.xctemplate/ "${TEMPLATE_DIR}"
+cp -r EngineerMemo.xctemplate/My\ MVVM\ Architecture "${TEMPLATE_DIR}"
